refactor(app): colocate route-local modules in underscore folders
Move Batch 1 route-local clients and search state helpers into `_components`/`_lib` and rewire imports to direct canonical paths, keeping behavior unchanged while making route structure easier to navigate and maintain.

Made-with: Cursor
