+++ src/app/draft-history/class-leaderboard/_components/ClassLeaderboardClient.tsx
